+++ src/builder.rs
@@ -15,6 +15,14 @@ impl JsRuntimeBuilder {
         }
     }
 
+    pub fn extensions(&mut self) -> &mut Vec<bjs::Extension> {
+        &mut self.options.extensions
+    }
+
+    pub fn resource_table(&mut self) -> &mut bjs::ResourceTable {
+        &mut self.options.resource_table
+    }
+
     // /// Whether inspector should break on the first line and wait
     // pub fn inspector_should_wait(mut self, inspector_should_wait: bool) -> Self {
     //     self.options.worker.should_break_on_first_statement = inspector_should_wait;
@@ -22,14 +30,20 @@ impl JsRuntimeBuilder {
     // }
 
     /// Sets the module loader to use to load `JavaScript`
-    pub fn with_module_loader(&mut self, loader: Rc<dyn bjs::ModuleLoader>) -> &mut Self {
+    pub fn with_module_loader(mut self, loader: Rc<dyn bjs::ModuleLoader>) -> Self {
         self.options.module_loader.replace(loader);
         self
     }
 
     /// Adds JavaScript [Extension](bjs::Extension)
-    pub fn with_extension(&mut self, extension: bjs::Extension) -> &mut Self {
+    pub fn with_extension(mut self, extension: bjs::Extension) -> Self {
         self.options.extensions.push(extension);
+        self
+    }
+
+    /// Adds Javascript [Resource](bjs::Resource)
+    pub fn with_resource(mut self, resource: Rc<dyn bjs::Resource>) -> Self {
+        self.options.resource_table.add_rc_dyn(resource);
         self
     }
