+++ src/runtimes/bevy/ext/bevy_ecs/mod.rs
@@ -16,7 +16,7 @@ pub fn init() -> Extension {
             "05_world.js",
         ))
         .ops(vec![
-            op_wait_for_world::decl(),
+            op_wait_for_frame::decl(),
             entity::op_entity_spawn::decl(),
             entity::op_entity_insert_component::decl(),
             entity::op_entity_get_component::decl(),
@@ -26,9 +26,10 @@ pub fn init() -> Extension {
         .build()
 }
 
+/// Waits until world becomes available to read from
 #[op]
-async fn op_wait_for_world(state: &mut OpState, rid: bjs::ResourceId) {
-    bjs::runtimes::unwrap_world_resource(state, rid)
-        .wait_for_world()
+async fn op_wait_for_frame(state: &mut OpState, world_resource_id: u32) {
+    bjs::runtimes::unwrap_world_resource(state, world_resource_id)
+        .wait_for_frame()
         .await
 }
